First version of my CEX-Tools.
CexSearch: Find which testcase(s) cover a given line. The line to be
searched must be given as signature-String of the method and offset
within method. Parses a specified directory tree for CEX-Files that
contains a COVERED P-Line with the specified offset under a B-Line with
the specified signature. 

cextoolbox: Read one or more complete CEX-Files into memory, manipulate them there and write a modified CEX to the file system.